cleanup: naming, moving things to shared
